+++ infra/lib/helper/security.ts
@@ -72,12 +72,12 @@ export function kmsKeyPolicyStatementGenerator(kmsKey?: kms.IKey): iam.PolicySta
     });
 }
 
-export function kmsKeyPolicyStatementPrincipalGenerator(kmsKey?: kms.IKey): iam.PolicyStatement {
+export function kmsKeyPolicyStatementPrincipalGenerator(config: Config.Config, kmsKey?: kms.IKey): iam.PolicyStatement {
     if (!kmsKey) {
         throw new Error("Cannot generate policy statement for KMS key if no KMS key provided.");
     }
 
-    return new iam.PolicyStatement({
+    const policyStatement =  new iam.PolicyStatement({
         actions: [
             "kms:GenerateDataKey*",
             "kms:Decrypt",
@@ -97,10 +97,24 @@ export function kmsKeyPolicyStatementPrincipalGenerator(kmsKey?: kms.IKey): iam.
             Service("ECS_TASKS").Principal,
             Service("LOGS").Principal,
             Service("LAMBDA").Principal,
-            Service("STS").Principal,
-            Service("CLOUDFRONT").Principal,
+            Service("STS").Principal
         ],
     });
+
+    if(!config.app.useAlb.enabled) {
+        policyStatement.addPrincipals(Service("CLOUDFRONT").Principal);
+    }
+
+    if(config.app.openSearch.useProvisioned.enabled) {
+        policyStatement.addPrincipals(Service("ES").Principal);
+    }
+
+    if(config.app.openSearch.useServerless.enabled) {
+        policyStatement.addPrincipals(Service("AOSS").Principal);
+    }
+
+
+    return policyStatement;
 }
 
 export function generateUniqueNameHash(
@@ -122,21 +136,26 @@ export function generateUniqueNameHash(
 export function generateContentSecurityPolicy(
     storageResources: storageResources,
     authenticationDomain: string,
-    apiUrl: string
+    apiUrl: string,
+    config: Config.Config
 ): string {
-    const connectSrc = [
+    let connectSrc = [
         "'self'",
         "blob:",
         authenticationDomain,
         `https://${Service("COGNITO_IDP").Endpoint}/`,
         `https://${Service("COGNITO_IDENTITY").Endpoint}/`,
-        `https://maps.${Service("GEO").Endpoint}/`,
         `https://${apiUrl}`,
         //`https://${props.storageResources.s3.assetBucket.bucketRegionalDomainName}/`, //Virtual Host Format Connection
         //`https://${props.storageResources.s3.assetBucket.bucketDomainName}/`, //Virtual Host Format Connection
         `https://${Service("S3").PrincipalString}/${storageResources.s3.assetBucket.bucketName}/`, //Path Addressable Format Connection
         `https://${Service("S3").Endpoint}/${storageResources.s3.assetBucket.bucketName}/`, //Path Addressable Format Connection
     ];
+
+    //Add GeoLocation service URL if feature turned on
+    if (config.app.useLocationService.enabled) {
+        connectSrc.push(`https://maps.${Service("GEO").Endpoint}/`);
+    }
 
     const scriptSrc = [
         "'self'",