+++ infra/lib/nestedStacks/auth/constructs/cognito-web-native-construct.ts
@@ -18,6 +18,7 @@ import { LAMBDA_PYTHON_RUNTIME } from "../../../../config/config";
 import { NagSuppressions } from "cdk-nag";
 import { Service } from "../../../helper/service-helper";
 import * as Config from "../../../../config/config";
+import { handler } from "../../search/constructs/schemaDeploy/provisioned/deployschemaprovisioned";
 
 export interface SamlSettings {
     metadata: cognito.UserPoolIdentityProviderSamlMetadata;
@@ -51,7 +52,8 @@ export class CognitoWebNativeConstructStack extends Construct {
     constructor(parent: Construct, name: string, props: CognitoWebNativeConstructStackProps) {
         super(parent, name);
 
-        const handlerName = "pretokengen";
+        //Check if GovCloud is enabled and set the handler to v1 instead (GovCloud does not support advanced security mode which can use the v2 pretokengen lambdas)
+        const handlerName = props.config.app.govCloud.enabled? "pretokengenv1" : "pretokengenv2";
         const preTokenGeneration = new lambda.Function(this, handlerName, {
             code: lambda.Code.fromAsset(path.join(__dirname, `../../../../../backend/backend`)),
             handler: `handlers.auth.${handlerName}.lambda_handler`,
@@ -101,18 +103,20 @@ export class CognitoWebNativeConstructStack extends Construct {
 
         const cfnUserPool = userPool.node.defaultChild as cognito.CfnUserPool;
 
-        //Add pretokengen lambda trigger (V2) - this will generate claims for both ID and Access token claims
+        //(Non-GovCloud) Add pretokengen lambda trigger (V2) - this will generate claims for both Access and ID token claims
+        //(GovCloud) Add pretokengen lambda trigger (V1) - this will generate claims for only Access token claims (ID token will not have claims and can't be used)
         cfnUserPool.lambdaConfig = {
             preTokenGenerationConfig: {
                 lambdaArn: preTokenGeneration.functionArn,
-                lambdaVersion: "V2_0",
+                lambdaVersion: props.config.app.govCloud.enabled? "V1_0" : "V2_0",
             },
         };
+
         userPool.node.addDependency(preTokenGeneration);
         preTokenGeneration.grantInvoke(Service("COGNITO_IDP").Principal);
 
+        //Only enable advanced security for non-govcloud environments (currently no supported by cognito)
         if (!props.config.app.govCloud.enabled) {
-            //Only enable advanced security where it's available
             const userPoolAddOnsProperty: cognito.CfnUserPool.UserPoolAddOnsProperty = {
                 advancedSecurityMode: "ENFORCED",
             };